Enum landlock::AccessFs

#[non_exhaustive]

#[repr(u64)]
pub enum AccessFs {

    Execute = 1,
    WriteFile = 2,
    ReadFile = 4,
    ReadDir = 8,
    RemoveDir = 16,
    RemoveFile = 32,
    MakeChar = 64,
    MakeDir = 128,
    MakeReg = 256,
    MakeSock = 512,
    MakeFifo = 1_024,
    MakeBlock = 2_048,
    MakeSym = 4_096,
    Refer = 8_192,
    Truncate = 16_384,

}

File system access right.

Each variant of AccessFs is an access right for the file system. A set of access rights can be created with BitFlags<AccessFs>.

Example

use landlock::{ABI, Access, AccessFs, BitFlags, make_bitflags};

let exec = AccessFs::Execute;

let exec_set: BitFlags<AccessFs> = exec.into();

let file_content = make_bitflags!(AccessFs::{Execute | WriteFile | ReadFile});

let fs_v1 = AccessFs::from_all(ABI::V1);

let without_exec = fs_v1 & !AccessFs::Execute;

assert_eq!(fs_v1 | AccessFs::Refer, AccessFs::from_all(ABI::V2));

Warning

To avoid unknown restrictions don’t use BitFlags::<AccessFs>::all() nor BitFlags::ALL, but use a version you tested and vetted instead, for instance AccessFs::from_all(ABI::V1). Direct use of the BitFlags API is deprecated. See ABI for the rationale and help to test it.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Execute = 1

Execute a file.

WriteFile = 2

Open a file with write access.

ReadFile = 4

Open a file with read access.

ReadDir = 8

Open a directory or list its content.

RemoveDir = 16

Remove an empty directory or rename one.

RemoveFile = 32

Unlink (or rename) a file.

MakeChar = 64

Create (or rename or link) a character device.

MakeDir = 128

Create (or rename) a directory.

MakeReg = 256

Create (or rename or link) a regular file.

MakeSock = 512

Create (or rename or link) a UNIX domain socket.

MakeFifo = 1_024

Create (or rename or link) a named pipe.

MakeBlock = 2_048

Create (or rename or link) a block device.

MakeSym = 4_096

Create (or rename or link) a symbolic link.

Refer = 8_192

Link or rename a file from or to a different directory.

Truncate = 16_384

Truncate a file with truncate(2), ftruncate(2), creat(2), or open(2) with O_TRUNC.

Implementations

impl AccessFs

pub fn from_read(abi: ABI) -> BitFlags<Self>

Gets the access rights identified as read-only according to a specific ABI. Exclusive with from_write().

pub fn from_write(abi: ABI) -> BitFlags<Self>

Gets the access rights identified as write-only according to a specific ABI. Exclusive with from_read().

pub fn from_file(abi: ABI) -> BitFlags<Self>

Gets the access rights legitimate for non-directory files.