landlock
Enum AccessNet

#[non_exhaustive]

#[repr(u64)]
pub enum AccessNet {
    BindTcp = 1,
    ConnectTcp = 2,
}

Network access right.

Each variant of AccessNet is an access right for the network. A set of access rights can be created with BitFlags<AccessNet>.

Example

use landlock::{ABI, Access, AccessNet, BitFlags, make_bitflags};

let bind = AccessNet::BindTcp;

let bind_set: BitFlags<AccessNet> = bind.into();

let bind_connect = make_bitflags!(AccessNet::{BindTcp | ConnectTcp});

let net_v4 = AccessNet::from_all(ABI::V4);

assert_eq!(bind_connect, net_v4);

Warning

To avoid unknown restrictions don’t use BitFlags::<AccessNet>::all() nor BitFlags::ALL, but use a version you tested and vetted instead, for instance AccessNet::from_all(ABI::V4). Direct use of the BitFlags API is deprecated. See ABI for the rationale and help to test it.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

BindTcp = 1

Bind to a TCP port.

ConnectTcp = 2

Connect to a TCP port.