Trait landlock::RulesetAttr

pub trait RulesetAttr: Sized + AsMut<Ruleset> + Compatible {
    // Provided method
    fn handle_access<T, U>(self, access: T) -> Result<Self, RulesetError>
       where T: Into<BitFlags<U>>,
             U: Access { ... }
}

Provided Methods

fn handle_access<T, U>(self, access: T) -> Result<Self, RulesetError>
where T: Into<BitFlags<U>>, U: Access,

Attempts to add a set of access rights that will be supported by this ruleset. By default, all actions requiring these access rights will be denied. Consecutive calls to handle_access() will be interpreted as logical ORs with the previous handled accesses.

On error, returns a wrapped HandleAccessesError. E.g., RulesetError::HandleAccesses(HandleAccessesError::Fs(HandleAccessError<AccessFs>))

Object Safety

This trait is not object safe.